Tested docker volume. edited user id to be big integer for docker.

+++ src/main/java/team2/hackathon/DatabaseSequence.java
@@ -4,6 +4,8 @@ import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigInteger;
+
 @Data
 @Document(collection = "database_sequences")
 public class DatabaseSequence {
@@ -11,7 +13,7 @@ public class DatabaseSequence {
     @Id
     private String id;
 
-    private long seq;
+    private BigInteger seq;
 
 
 }
